App: run an http server instead of stdout
I had wanted to keep it really simple (metrics on stdout) for the sake
of reduced memory usage. But it turns out, since that required another
service to serve the metrics it wasn't less complicated and the
persistent node.js process only takes ~50MB of RAM.

Start an http server, cache the alertra data for some time (because they
don't change often), and generate metrics for every request (ignore url).

This also sends a 500 on error and logs the message to stderr.