Graphite: add support for reporting metrics
Graphite works as a push system so if the env vars are set we use
an interval to connect and write the metics to graphite.

We write the same three metrics as with prometheus with the same set of
labels (called tags in graphite).

Connection errors are printed to stderr.