App: explicitly add signal handlers for Docker
Docker runs the proc as PID 1, which is special and doesn't
automatically handle signals and terminate... or something.

We either need to tell docker to run *another* proc as PID 1 that
forwards signals, or explicitly adds handlers. Adding handlers is easy,
so let's do that.

Now it dies quickly when docker or Ctrl-C asks it to.